requirements.txt

appdirs==1.4.3
backcall==0.1.0
beautifulsoup4==4.7.1
btrfsutil==1.1.0
CacheControl==0.12.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cupshelpers==1.0
decorator==4.4.0
distlib==0.2.8
distro==1.4.0
dnspython==1.16.0
docopt==0.6.2
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.2.1
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
ipython==7.5.0
ipython-genutils==0.1.0
itsdangerous==1.1.0
Jade-Application-Kit==1.3.5
jedi==0.13.3
Jinja2==2.10.1
keyutils==0.5
LibAppArmor==2.13.2
lockfile==0.12.2
louis==3.9.0
lxml==4.3.3
MarkupSafe==1.1.1
monotonic==1.5
msgpack==0.6.1
mysql-connector-python==8.0.15
npyscreen==4.10.5
packaging==19.0
pacman-mirrors==4.14.2
parso==0.4.0
pep517==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
progress==1.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pwquality==1.4.0
pycairo==1.18.0
pycups==1.9.74
pycurl==7.43.0.2
Pygments==2.4.0
PyGObject==3.32.1
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
pysmbc==1.0.15.8
python-engineio==3.3.1
python-libtorrent==1.1.12
python-socketio==3.1.2
pytoml==0.1.20
pyxdg==0.26
PyYAML==3.13
reportlab==3.5.13
requests==2.21.0
retrying==1.3.3
simplegeneric==0.8.1
six==1.12.0
soupsieve==1.9.1
team==1.0
traitlets==4.3.2
udiskie==1.7.7
urllib3==1.25.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
websockets==7.0
Werkzeug==0.15.2
